+++ backend/src/main/java/nulp/cs/carrentalrestservice/modules/payment/dto/StripeRequest.java
@@ -5,12 +5,14 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
 public class StripeRequest {
-    private long amount;
+    private BigDecimal amount;
     private String currency;
     private String name;
 }
